bannertopdf filter converted to filter function
